Add support to expose functions on a public domain
