+++ lib/domain/storage/redis.js
@@ -210,6 +210,8 @@ class StorageRedis extends Storage {
 
     if (code.exposed === true) {
       data.exposed = 'true';
+    } else if (code.exposed === false) {
+      data.exposed = 'false';
     }
 
     this.setNamespaceMember(namespace, id);
