+++ lib/http/routers/FunctionsRouter.js
@@ -4,14 +4,14 @@ const Router = require('express').Router;
 const bodyParser = require('body-parser');
 const Validator = require('jsonschema').Validator;
 
+const functionRunHandler = require('./functionRunHandler');
+const FunctionsRequest = require('../FunctionsRequest');
+const SchemaResponse = require('../SchemaResponse');
+
 const log = require('../../support/log');
 const schemas = require('../../domain/schemas');
 const Pipeline = require('../../domain/Pipeline');
-const ErrorTracker = require('../../domain/ErrorTracker');
-const { StdoutLogStorage, DefaultLogStorage } = require('../../domain/LogStorage');
-const FunctionsRequest = require('../FunctionsRequest');
-const Metric = require('../../domain/Metric');
-const SchemaResponse = require('../SchemaResponse');
+const { StdoutLogStorage } = require('../../domain/LogStorage');
 
 const router = new Router();
 const { bodyParserLimit } = require('../../support/config');
@@ -140,7 +140,7 @@ router.put('/:namespace/:id', bodyParser.json({ limit: bodyParserLimit }), async
     data.env = env;
   }
 
-  if (exposed) {
+  if (exposed !== undefined) {
     data.exposed = exposed;
   }
 
@@ -246,85 +246,11 @@ router.delete('/:namespace/:id', async (req, res) => {
 });
 
 
-router.all('/:namespace/:id/run', bodyParser.json({ limit: bodyParserLimit }), async (req, res) => {
-  const { namespace, id } = req.params;
-  const memoryStorage = req.app.get('memoryStorage');
-  const sandbox = req.app.get('sandbox');
-  const filename = codeFileName(namespace, id);
-  const metric = new Metric('function-run');
-  const logStorage = new DefaultLogStorage(namespace, id, req);
-
-  let code;
-
-  try {
-    code = await memoryStorage.getCodeByCache(namespace, id, {
-      preCache: (preCode) => {
-        preCode.script = sandbox.compileCode(filename, preCode.code);
-        return preCode;
-      },
-    });
-
-    if (!code) {
-      const error = new Error(`Code '${namespace}/${id}' is not found`);
-      error.statusCode = 404;
-      throw error;
-    }
-  } catch (err) {
-    res.status(err.statusCode || 500).json({ error: err.message });
-    return;
-  }
-
-  try {
-    const options = {
-      console: logStorage.console,
-      env: code.env,
-    };
-    const result = await sandbox.runScript(code.script, req, options);
-
-    res.set(result.headers);
-    res.status(result.status);
-    res.json(result.body);
-
-    const spent = metric.finish({
-      filename,
-      status: result.status,
-    });
-
-    logStorage.flush({
-      status: result.status,
-      requestTime: spent,
-    });
-  } catch (err) {
-    logStorage.console.error(`Failed to run function: ${err}`);
-    logStorage.console.error(err.stack);
-    const status = err.statusCode || 500;
-    res.status(status).json({ error: err.message });
-
-    const spent = metric.finish({
-      filename,
-      status,
-      error: err.message,
-    });
-
-    const logResult = logStorage.flush({
-      status,
-      requestTime: spent,
-    });
-
-    const { namespaceSettings } = code;
-    const { sentryDSN } = namespaceSettings || {};
-
-    const extra = Object.assign({ body: req.body }, logResult || {});
-    const errTracker = new ErrorTracker({
-      sentryDSN,
-      filename,
-      extra,
-      tags: { codeHash: code.hash },
-      code: code.code,
-    });
-    errTracker.notify(err);
-  }
-});
+router.all(
+  '/:namespace/:id/run',
+  bodyParser.json({ limit: bodyParserLimit }),
+  (req, res) => functionRunHandler(req, res, { exposed: false })
+);
 
 
 router.put('/pipeline', bodyParser.json({ limit: bodyParserLimit }), async (req, res) => {
