+++ lib/http/routes.js
@@ -1,4 +1,5 @@
 const express = require('express');
+const vhost = require('vhost');
 const morgan = require('morgan');
 
 const Sandbox = require('@globocom/backstage-functions-sandbox');
@@ -10,6 +11,7 @@ const StatusRouter = require('./routers/StatusRouter');
 const DebugRouter = require('./routers/DebugRouter');
 const NamespacesRouter = require('./routers/NamespacesRouter');
 const FunctionsRouter = require('./routers/FunctionsRouter');
+const FunctionsPublicRouter = require('./routers/FunctionsPublicRouter');
 const RedisStorage = require('../domain/storage/redis');
 const parseExposeEnv = require('../support/parseExposeEnv');
 const config = require('../support/config');
@@ -26,19 +28,24 @@ morgan.token('response-sectime', (req, res) => {
   return secs.toFixed(3);
 });
 
-const app = express();
-
-app.use(morgan(config.log.morganFormat));
-app.disable('x-powered-by');
-app.enable('trust proxy');
-
-app.set('memoryStorage', new RedisStorage());
-app.set('sandbox', new Sandbox({
+const memoryStorage = new RedisStorage();
+const sandbox = new Sandbox({
   env: parseExposeEnv(),
   globalModules: config.defaultGlobalModules,
   asyncTimeout: config.asyncTimeout,
   syncTimeout: config.syncTimeout,
-}));
+});
+
+function setupApp(app) {
+  app.use(morgan(config.log.morganFormat));
+  app.disable('x-powered-by');
+  app.enable('trust proxy');
+  app.set('memoryStorage', memoryStorage);
+  app.set('sandbox', sandbox);
+}
+
+const app = express();
+setupApp(app);
 
 app.get('/', (req, res) => {
   const backstageRequest = new FunctionsRequest(req);
@@ -48,6 +55,17 @@ app.get('/', (req, res) => {
       name: 'Backstage functions',
     });
 });
+
+if (config.exposed.host) {
+  const publicApp = express();
+  setupApp(publicApp);
+  publicApp.use(FunctionsPublicRouter);
+  publicApp.use((req, res) => {
+    res.send({ error: 'Not found' });
+  });
+
+  app.use(vhost(config.exposed.host, publicApp));
+}
 
 app.use('/healthcheck', HealthcheckRouter);
 app.use('/status', StatusRouter);
