+++ lib/support/config.js
@@ -37,4 +37,7 @@ module.exports = {
     useCertFile: ConfigDiscovery.getBool('FUNCTIONS_USE_SSL_CERT_FILE', false),
     certFile: process.env.SSL_CERT_FILE,
   },
+  exposed: {
+    host: process.env.EXPOSED_HOST,
+  },
 };
